% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-client.R
\name{bd_request}
\alias{bd_request}
\title{Abstraction of an http request to the API}
\usage{
bd_request(endpoint, query = list())
}
\arguments{
\item{endpoint}{which endpoint to request.}

\item{query}{querystring passed as argument to the GEt request.}
}
\value{
A httr::response() object.
}
\description{
Abstraction of an http request to the API
}
\examples{
\dontrun{

bd_request(
 endpoint = "dataset_search",
 query = list(
  resource_type = "bdm_table",
  q = search_term,
  page_size = 100))

bd_request(
 endpoint = "bdm_dataset_show",
 query = list(
  dataset_id = dataset_id)

}
}
